% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocks.R
\name{create_vaccine}
\alias{create_vaccine}
\title{A single vaccine dose, as a fragment}
\usage{
create_vaccine(label, code, series = 1)
}
\arguments{
\item{label}{The state's name.}

\item{code}{A \code{create_component_settings("code", ...)} result.}

\item{series}{The dose number (an integer). Defaults to \code{1}.}
}
\value{
A fragment.
}
\description{
A multi-dose series is \code{chain()}ing several of these with \code{create_delay()} between them -- no
separate "series" construct needed.
}
\examples{
bnt162b2 <- create_component_settings("code",
  system = "CVX", code = "208",
  display = "COVID-19 vaccine, mRNA"
)
create_vaccine("First Dose", bnt162b2, series = 1)
}
