% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{default_recurring_drugs}
\alias{default_recurring_drugs}
\title{Default list of drug concepts given as repeat prescriptions}
\usage{
default_recurring_drugs()
}
\value{
A character vector of \verb{DP_*} concept ids.
}
\description{
Long-term therapies (antihypertensives, statins, anticoagulants, immunosuppressants,
psychotropics, ...) from the RWE-BRIDGE drug-proxy codelist. Every other drug concept defaults
to a single short \code{course}. Edit \code{concepts.csv} afterwards to change any of them.
}
\examples{
default_recurring_drugs()
}
