% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheaModuleR-package.R
\docType{package}
\name{syntheaModuleR-package}
\alias{syntheaModuleR}
\alias{syntheaModuleR-package}
\title{syntheaModuleR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UMC-Utrecht-RWE/syntheaModuleR}
  \item Report bugs at \url{https://github.com/UMC-Utrecht-RWE/syntheaModuleR/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
